I dont need sleep, I need answerz

+++ src/main/java/sk/tuke/kpi/oop/game/actions/PerpetualReactorHeating.java
@@ -15,8 +15,8 @@ public class PerpetualReactorHeating extends AbstractAction<Reactor>{
         }
         this.reactor = reactor;
     }
-    public PerpetualReactorHeating(float temperature){
-        this.number = (int) temperature;
+    public PerpetualReactorHeating(int temperature){
+        this.number = temperature;
     }
 
     @Override
